Fix an absolute metric load of possible bugs and IDEA code suggestions

+++ src/main/java/zmaster587/advancedRocketry/client/render/planet/RenderSpaceSky.java
@@ -26,8 +26,6 @@ import zmaster587.libVulpes.util.Vector3F;
 
 public class RenderSpaceSky extends RenderPlanetarySky {
 
-	//Mostly vanilla code
-	//TODO: make usable on other planets
 	public RenderSpaceSky() {
 		super();
 	}
@@ -42,8 +40,6 @@ public class RenderSpaceSky extends RenderPlanetarySky {
 	
 	@Override
 	public void renderPlanet2(BufferBuilder buffer, MatrixStack matrix, DimensionProperties properties, float size, float alphaMultiplier, double shadowAngle, boolean hasRing, float[] shadowColorMultiplier, float alphaMultiplier2) {
-		//ResourceLocation icon, int locationX, int locationY, double zLevel, float planetOrbitalDistance, float alphaMultiplier, double angle, boolean hasAtmosphere, float[] atmColor, float[] ringColor, boolean isGasgiant, boolean hasRings, boolean hasDecorators) {
-
 		BlockPos playerPos = new BlockPos(mc.player.getPositionVec());
 		ISpaceObject object = SpaceObjectManager.getSpaceManager().getSpaceStationFromBlockCoords(playerPos);
 
@@ -51,8 +47,7 @@ public class RenderSpaceSky extends RenderPlanetarySky {
 			return;
 		float planetOrbitalDistance = object.getOrbitalDistance();
 		
-		if(properties.isStar())
-		{
+		if(properties.isStar()) {
 			size = 10;
 			RenderSystem.depthMask(true);
 			RenderSystem.enableAlphaTest();
@@ -76,16 +71,15 @@ public class RenderSpaceSky extends RenderPlanetarySky {
 			buffer.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX);	
 			f10 = size*2f*AstronomicalBodyHelper.getBodySizeMultiplier(planetOrbitalDistance);
 			//multiplier = 2;
-			RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)(-f10)).tex(0.0f, 0.0f).endVertex();
-			RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)(-f10)).tex(1.0f, 0.0f).endVertex();
-			RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)f10).tex(1.0f, 1.0f).endVertex();
-			RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)f10).tex(0.0f, 1.0f).endVertex();
+			RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, -f10).tex(0.0f, 0.0f).endVertex();
+			RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, -f10).tex(1.0f, 0.0f).endVertex();
+			RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, f10).tex(1.0f, 1.0f).endVertex();
+			RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, f10).tex(0.0f, 1.0f).endVertex();
 			Tessellator.getInstance().draw();
 			matrix.pop();
 			
 			GL11.glDepthMask(false);
-			for(int i = 0; i < 3; i++)
-			{
+			for(int i = 0; i < 3; i++) {
 				float speedMult = (i)*1.01f + 1;
 				//Render accretion disk
 				mc.getTextureManager().bindTexture(TextureResources.locationAccretionDisk);
@@ -97,10 +91,10 @@ public class RenderSpaceSky extends RenderPlanetarySky {
 				RenderSystem.color4f((float)1, (float).5 , (float).4 ,1f);
 				buffer.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX);	
 				f10 = size*40f*AstronomicalBodyHelper.getBodySizeMultiplier(planetOrbitalDistance);
-				RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)(-f10)).tex(0.0f, 0.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)(-f10)).tex(1.0f, 0.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)f10).tex(1.0f, 1.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)f10).tex(0.0f, 1.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, -f10).tex(0.0f, 0.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, -f10).tex(1.0f, 0.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, f10).tex(1.0f, 1.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, f10).tex(0.0f, 1.0f).endVertex();
 				Tessellator.getInstance().draw();
 				matrix.pop();
 				
@@ -114,10 +108,10 @@ public class RenderSpaceSky extends RenderPlanetarySky {
 				buffer.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX);	
 				f10 = size*30f*AstronomicalBodyHelper.getBodySizeMultiplier(planetOrbitalDistance);
 				//multiplier = 2;
-				RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)(-f10)).tex(0.0f, 0.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)(-f10)).tex(1.0f, 0.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)f10).tex(1.0f, 1.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)f10).tex(0.0f, 1.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, -f10).tex(0.0f, 0.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, -f10).tex(1.0f, 0.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, f10).tex(1.0f, 1.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, f10).tex(0.0f, 1.0f).endVertex();
 				Tessellator.getInstance().draw();
 				matrix.pop();
 				
@@ -131,10 +125,10 @@ public class RenderSpaceSky extends RenderPlanetarySky {
 				buffer.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX);	
 				f10 = size*15f*AstronomicalBodyHelper.getBodySizeMultiplier(planetOrbitalDistance);
 				//multiplier = 2;
-				RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)(-f10)).tex(0.0f, 0.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)(-f10)).tex(1.0f, 0.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)f10, 0.0D, (double)f10).tex(1.0f, 1.0f).endVertex();
-				RenderHelper.vertexPos(matrix, buffer, (double)(-f10), 0.0D, (double)f10).tex(0.0f, 1.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, -f10).tex(0.0f, 0.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, -f10).tex(1.0f, 0.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, f10, 0.0D, f10).tex(1.0f, 1.0f).endVertex();
+				RenderHelper.vertexPos(matrix, buffer, -f10, 0.0D, f10).tex(0.0f, 1.0f).endVertex();
 				Tessellator.getInstance().draw();
 				matrix.pop();
 			}
@@ -178,10 +172,10 @@ public class RenderSpaceSky extends RenderPlanetarySky {
 		RenderSystem.color4f(1f, 1f, 1f, alphaMultiplier);
 
 		buffer.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION_TEX);
-		RenderHelper.vertexPos(matrix, buffer, (double)(-f10), -10.0D, (double)f10).tex(f16, f17).endVertex();
-		RenderHelper.vertexPos(matrix, buffer, (double)f10, -10.0D, (double)f10).tex(f14, f17).endVertex();
-		RenderHelper.vertexPos(matrix, buffer, (double)f10, -10.0D, (double)(-f10)).tex(f14, f15).endVertex();
-		RenderHelper.vertexPos(matrix, buffer, (double)(-f10), -10.0D, (double)(-f10)).tex(f16, f15).endVertex();
+		RenderHelper.vertexPos(matrix, buffer, -f10, -10.0D, f10).tex(f16, f17).endVertex();
+		RenderHelper.vertexPos(matrix, buffer, f10, -10.0D, f10).tex(f14, f17).endVertex();
+		RenderHelper.vertexPos(matrix, buffer, f10, -10.0D, -f10).tex(f14, f15).endVertex();
+		RenderHelper.vertexPos(matrix, buffer, -f10, -10.0D, -f10).tex(f16, f15).endVertex();
 
 		Tessellator.getInstance().draw();
 		GL11.glPopAttrib();
