+++ src/main/java/zmaster587/advancedRocketry/tile/multiblock/energy/TileSolarArray.java
@@ -1,6 +1,5 @@
 package zmaster587.advancedRocketry.tile.multiblock.energy;
 
-import io.netty.buffer.ByteBuf;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.Blocks;
 import net.minecraft.entity.player.PlayerEntity;
@@ -145,7 +144,7 @@ public class TileSolarArray extends TileMultiPowerProducer implements ITickableT
 				//Slight adjustment to make Earth 0.9995 into a 1.0
 				energyRecieved = (int) (numPanels * 1.0005d * 2 * insolationPowerMultiplier);
 			}
-			powerMadeLastTick = (int) (energyRecieved*ARConfiguration.getCurrentConfig().solarGeneratorMult.get());
+			powerMadeLastTick = energyRecieved*ARConfiguration.getCurrentConfig().solarGeneratorMult.get();
 
 			if(powerMadeLastTick != prevPowerMadeLastTick) {
 				prevPowerMadeLastTick = powerMadeLastTick;
